refactored a bit and have models in seperate file, still cannot join the two tables, how does relationship really work?
