+++ backend/main.py
@@ -2,51 +2,15 @@ from flask import Flask, request
 from flask_restful import Api, Resource, reqparse, abort, fields, marshal_with
 from flask_sqlalchemy import SQLAlchemy
 import sqlite3
+from models import VideoModel, CommentModel, db
 from random import randrange
 
 app = Flask(__name__)
 api = Api(app)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
-db = SQLAlchemy(app)
+db.init_app(app)
 
-class VideoModel(db.Model):
-    __tablename__ = 'video'
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(100), nullable=False)
-    views = db.Column(db.Integer, nullable=False)
-    likes = db.Column(db.Integer, nullable=False)
-    comments = db.relationship(
-        'CommentModel', 
-        backref='video',
-        cascade='all, delete, delete-orphan',
-        single_parent=True,
-        lazy=True
-    ) 
-
-    def __init__(self, id=None, name=None, views=None, likes=None):
-        self.id = id
-        self.name = name
-        self.views = views
-        self.likes = likes
-
-    def __repr__(self):
-        return f"video(name={name}, views={views}, likes={likes}, comments={comments})"
-
-class CommentModel(db.Model):
-    __tablename__ = 'comment'
-    id = db.Column(db.Integer, primary_key=True)
-    content = db.Column(db.String(1000), nullable=False)
-    video_id = db.Column(db.Integer, db.ForeignKey('video.id'), nullable=False)
-
-    def __init__(self, id=None, content=None, video_id=None):
-        self.id = id
-        self.content = content
-        self.video_id = video_id
-
-    def __repr__(self):
-        return f"comment(content={content}, video_id={video_id})"
-
-db.create_all()
+#db.create_all()
 
 video_put_args = reqparse.RequestParser()
 video_put_args.add_argument("name", type=str, help="Name of the video is required", required=True)
@@ -71,7 +35,7 @@ resource_fields = {
 class Video(Resource):
     @marshal_with(resource_fields)
     def get(self, video_id): 
-        result = VideoModel.query.filter_by(id=video_id).first()
+        result = VideoModel.query.join(VideoModel.comment, aliased=True).filter_by(video_id=video_id)
         if not result:
             abort(404, message="Could not find video with that id")
         return result
@@ -127,9 +91,9 @@ resource_fields_comment = {
 class Comment(Resource):
     @marshal_with(resource_fields_comment)
     def get(self, video_id):
-        result = VideoModel.query.filter_by(id=video_id).first()
+        result = CommentModel.query.filter_by(video_id=video_id).first()
         if not result:
-            abort(404, message="Could not find video with that id")
+            abort(404, message="Could not find comments for that video id")
         return result
 
     @marshal_with(resource_fields_comment)
