+++ backend/readme.md
@@ -21,5 +21,5 @@ DELETE: curl -X DELETE http://127.0.0.1:5000/video/20
 
 **Comment requests**
 
-GET: curl -X GET http://127.0.0.1:5000/video/21
+GET: curl -X GET http://127.0.0.1:5000/video/21/comment
 PUT: curl -X PUT -H "Content-Type: application/json" --data '{"content": "this video is ok"}' http://127.0.0.1:5000/video/21/comment
